src/requirements.txt
azure-functions
tiktoken==0.6.0
azure-ai-documentintelligence==1.0.0b1